requirements.txt
flask
flask-cors
pyttsx3
pillow
